Added scripts deepFileGrep.sh and tarG
deepFileGrep.sh: This script allows you to perform a search for a given string in any file whose name contains your given search string.

tarG: This script unpacks tar files to a directory that is equivalent to the original file name with _unpack appended, and white space converted to underscores.